Rename _battery_pool_status.py → _component_pool_status_tracker.py
The renaming of the file is done in a separate commit so that the
detailed diff available in the previous commit doesn't get lost.
