+++ src/frequenz/sdk/actor/power_distributing/__init__.py
@@ -10,7 +10,7 @@ It also provides all the secondary features that should be used to communicate w
 PowerDistributingActor and send requests for charging or discharging power.
 """
 
-from ._battery_pool_status import ComponentStatus
+from ._component_pool_status_tracker import ComponentStatus
 from .power_distributing import PowerDistributingActor
 from .request import Request
 from .result import Error, OutOfBounds, PartialFailure, Result, Success
