+++ src/frequenz/sdk/actor/power_distributing/_component_pool_status_tracker.py
